add stats on StatsCommand
